requirements.txt
djangorestframework==3.11.0
pycryptodome==3.9.7
Pyrebase==3.0.27
pycrypto==2.6.1